✨ Type-aware integer wrapping removes manual masking from SHA-256
Integer literals are now untyped ("int", arbitrary precision). Fixed-width
types (u32, i32, u64, etc.) automatically wrap arithmetic results to their
range via mk_int/wrap_int. Wrapping happens at storage points: typed
variable declarations, const definitions with type annotations, and
ArrayValue.set() with element_type. This eliminates all & 4294967295 masks
from test_start.nl while producing identical SHA-256 output.

Key changes:
- value.py: resolve_width, wrap_int, coerce_to_type; ArrayValue tracks
  element_type and coerces on set()
- eval.py: all operators use resolve_width; VarDef/const coerce to
  declared types; rotate/shift derive bit width from operand types
- ast.py/lexer.py: integer literal default changed from i64 to int
- parser.py/main.py: const_assign carries type annotation
- newlang.md: document integer types and untyped int concept

+++ interp/ast.py
@@ -8,9 +8,9 @@ a tree of these nodes; the evaluator walks the tree to produce results.
 class IntLit:
     """Integer literal with an optional bit-width suffix."""
 
-    def __init__(self, value: int, width: str = "i64"):
+    def __init__(self, value: int, width: str = "int"):
         self.value = value
-        self.width = width  # e.g. "i8", "u32", "i64"
+        self.width = width
 
 
 class StrLit:
